Avoid creating spans around Stackdriver submit (#560)
The Stackdriver exporter uses a gRPC client that's instrumented with
OpenCensus. This means that there's normally a chance that traces will
be started from the span submission method. In the worst case, with
AlwaysSample, this adds an unbounded number of useless Spans.

This PR follows @bogdandrutu's advice and wraps the upload method in
a Span with NeverSample.

Fixes: #557