Do not use zero defaults for bundle delays (#34)
We were previously not checking for unset (zero) values in Options
before overriding the bundle delays in the bundler for the stats
portion of the exporter. Change this to only set the delays if
the values in the Options struct are non-zero, similar to what we
do for the trace exporter.