OpenCensusProto Point to Stackdriver Point
An adapter to directly convert from OpenCensus Proto
Point to Stackdriver's monitoring/v3 Point which are
both protobuf based code backed representations.

This the first of a series of changes to allow direct
conversion of metrics so that we can use this exporter
in the OpenCensus Agent/Service exporters, but also so that
applications such as OpenCensus-PHP's daemon can begin to it too.

Also added a TODO for SummaryValue conversion:
A reminder that we should support the conversion
from metricspb.SummaryValue to monitoring/v3 proto
whenever #66 is solved.

Updates #64.