Use standard label name for location. (#241)
* Use standard label name for location.

The location label is required. The resource type mappings defined in resource.go uniformly expect resourcekeys.CloudKeyZone, except the `generic_task`, but there is no good reason for that, as Stackdriver still requires a valid cloud location for `generic_task`. Therefore, use a cloud location everywhere.

This makes the exporter work out-of-the-box in conjuction with `kubernetes-operator` and partially addresses https://github.com/census-ecosystem/kubernetes-operator/issues/14

* Update stackdriver.go

Added printfs that are useful for diagnosing resource discovery.