trace: allow for concurrent uploads to Stackdriver
In situations where a large number of spans need to be exported from a
single instance of the exporter, bundle uploads are limited to using a
single goroutine. This limits the overall throughput of the exporter.

Make use of the NumberOfWorkers option to allow the exporter to use
multiple, concurrent goroutines to upload spans to Stackdriver.

Closes #245.

Signed-off-by: Nick Travers <n.e.travers@gmail.com>