Use generic_task resource type when all generic labels are present
While converting an ocagent resource to a cloud resource, stackdriver
exporter looks for a set of labels that apply to a generic_task resource:
https://cloud.google.com/monitoring/api/resources#tag_generic_task

However, if it doesn't find one of the labels, it falls back to "global"
resource with just the project id label, and it works OK.

But if it does find all of the labels, it will attach them to the exported
resource, and Cloud API fails with "Unrecognized resource label" error,
because those labels do not exist on a "global" resource.

This change replaces use of global resource with generic resource
throughout stackdriver exporter code.